refactor(pairing): extract leaf record types to break store/domain module cycle
