+++ src/infra/device-pairing.types.ts
@@ -1,0 +1,143 @@
+// Persisted device pairing and bootstrap-token record shapes.
+// Leaf contract shared by the domain modules (device-pairing.ts,
+// device-bootstrap.ts) and the SQLite row mapper (device-pairing-store.ts);
+// keeping it import-free of both sides prevents module cycles.
+import type { DeviceBootstrapProfile } from "../shared/device-bootstrap-profile.js";
+
+/** Pending device pairing request awaiting owner approval. */
+export type DevicePairingPendingRequest = {
+  requestId: string;
+  deviceId: string;
+  publicKey: string;
+  displayName?: string;
+  platform?: string;
+  deviceFamily?: string;
+  clientId?: string;
+  clientMode?: string;
+  role?: string;
+  roles?: string[];
+  scopes?: string[];
+  remoteIp?: string;
+  silent?: boolean;
+  isRepair?: boolean;
+  ts: number;
+};
+
+// Internal pending record. refreshedAtMs is a TTL keepalive stamped on refresh so an
+// actively retrying device keeps one pending request (and requestId) alive instead of
+// minting a new request every TTL window and flooding operator approval UIs. It never
+// crosses the protocol boundary, and ordering/--latest still use ts.
+export type DevicePairingPendingRecord = DevicePairingPendingRequest & {
+  refreshedAtMs?: number;
+};
+
+/** Bearer token issued to one paired device role. */
+export type DeviceAuthToken = {
+  token: string;
+  role: string;
+  scopes: string[];
+  issuer?: {
+    kind: "shared-gateway-auth";
+    generation: string;
+  };
+  createdAtMs: number;
+  rotatedAtMs?: number;
+  revokedAtMs?: number;
+  lastUsedAtMs?: number;
+};
+
+/**
+ * How the latest pairing approval was granted. "silent" is a same-host local
+ * policy approval and the only prune-eligible kind: local clients re-pair
+ * silently and cannot collide with another machine's records. "trusted-cidr"
+ * is also non-interactive but crosses hosts, so it is never pruned
+ * automatically (display metadata is not a machine identity). "owner" and
+ * "bootstrap" approvals required a user action and are never pruned.
+ */
+export type PairedDeviceApprovalKind = "owner" | "silent" | "trusted-cidr" | "bootstrap";
+
+/**
+ * Approved node capability surface for a node-role device. Device pairing
+ * grants connection auth; this grants command/capability exposure (node
+ * command gating). displayName here is the operator-facing node name set at
+ * approval or via node.rename; it must not be clobbered by reconnect
+ * metadata refreshes, which is why it lives apart from the device fields.
+ */
+export type PairedDeviceNodeSurface = {
+  displayName?: string;
+  version?: string;
+  coreVersion?: string;
+  uiVersion?: string;
+  modelIdentifier?: string;
+  caps?: string[];
+  commands?: string[];
+  permissions?: Record<string, boolean>;
+  bins?: string[];
+  createdAtMs: number;
+  approvedAtMs: number;
+  lastConnectedAtMs?: number;
+};
+
+/**
+ * Pending node-surface approval awaiting an operator decision (one per
+ * device). Carries its own metadata snapshot so approval UIs can show what
+ * the node declared at request time. `revision` guards the reconnect-vs-
+ * approve race: reconnect cleanup only deletes the revision it observed, so
+ * a refreshed request survives concurrent approval flows.
+ */
+export type PairedDevicePendingNodeSurface = {
+  requestId: string;
+  revision: string;
+  displayName?: string;
+  platform?: string;
+  version?: string;
+  coreVersion?: string;
+  uiVersion?: string;
+  clientId?: string;
+  clientMode?: string;
+  deviceFamily?: string;
+  modelIdentifier?: string;
+  caps?: string[];
+  commands?: string[];
+  permissions?: Record<string, boolean>;
+  remoteIp?: string;
+  silent?: boolean;
+  ts: number;
+};
+
+/** Persisted approved device record, including durable approval and active role tokens. */
+export type PairedDevice = {
+  deviceId: string;
+  publicKey: string;
+  displayName?: string;
+  platform?: string;
+  deviceFamily?: string;
+  clientId?: string;
+  clientMode?: string;
+  role?: string;
+  roles?: string[];
+  scopes?: string[];
+  approvedScopes?: string[];
+  remoteIp?: string;
+  tokens?: Record<string, DeviceAuthToken>;
+  approvedVia?: PairedDeviceApprovalKind;
+  nodeSurface?: PairedDeviceNodeSurface;
+  pendingNodeSurface?: PairedDevicePendingNodeSurface;
+  createdAtMs: number;
+  approvedAtMs: number;
+  lastSeenAtMs?: number;
+  lastSeenReason?: string;
+};
+
+/** Persisted bootstrap token state, including binding and role/scope redemption progress. */
+export type DeviceBootstrapTokenRecord = {
+  token: string;
+  ts: number;
+  deviceId?: string;
+  publicKey?: string;
+  profile?: DeviceBootstrapProfile;
+  redeemedProfile?: DeviceBootstrapProfile;
+  pendingProfile?: DeviceBootstrapProfile;
+  issuedAtMs: number;
+  lastUsedAtMs?: number;
+};
